Fix u16 overflow in CAP.MQES that crashed dma_alloc_coherent
CAP.MQES is a 0's-based field where the legal max value 0xffff means
65536 entries.  rc_nvme_init_controller computes mqes = MQES + 1, which
overflows when stored as u16 and reports mqes=0 in dmesg.  The
admin_sq_depth clamp then drives the requested depth to 0, and
dma_alloc_coherent(dev, 0, ...) trips a NULL deref deep inside
__iommu_dma_alloc_noncontiguous on kernels with AMD IOMMU active.

Widen mqes to u32 and update the CAP macro cast so the +1 doesn't
truncate before assignment.  Verified on TRX50 1022:B000 reporting
CAP=0x080000203c01ffff: MQES now logs as 65536, admin queue alloc
succeeds, ASQ/ACQ register read-back matches, controller reaches
CSTS.RDY=1.

+++ rc_linux.h
@@ -208,7 +208,7 @@ struct rc_nvme_state {
 
     // Capability snapshot (CAP at BAR0 + 0x00)
     u64            cap;
-    u16            mqes;              // CAP.MQES + 1 = max queue depth
+    u32            mqes;              // CAP.MQES + 1 = max queue depth (must be wider than u16: spec max is 65536)
     u8             dstrd;             // CAP.DSTRD doorbell stride
     u8             timeout_500ms;     // CAP.TO * 500 ms = boot timeout
 
@@ -565,7 +565,7 @@ int rc_raid_array_ioctl(struct block_device *bdev, fmode_t mode, unsigned int cm
 #define RC_NVME_REG_DBL_BASE		0x1000	/* Doorbell registers begin here */
 
 /* CAP field accessors (64-bit register) */
-#define RC_NVME_CAP_MQES(cap)		((u16)((cap) & 0xffff))		/* Max queue entries supported - 1 */
+#define RC_NVME_CAP_MQES(cap)		((u32)((cap) & 0xffff))		/* Max queue entries supported - 1 (0's based, so +1 needs u32) */
 #define RC_NVME_CAP_TO(cap)		((u8)(((cap) >> 24) & 0xff))	/* Timeout (500 ms units) */
 #define RC_NVME_CAP_DSTRD(cap)		((u8)(((cap) >> 32) & 0xf))	/* Doorbell stride */
 #define RC_NVME_CAP_CSS(cap)		((u8)(((cap) >> 37) & 0xff))	/* Command sets supported */
